backend/requirements.txt
django>=5.0
djangorestframework>=3.15
python-decouple>=3.8
pymysql>=1.1
